refactor sync connector
